General implementation of Revision indexing (including study variables) is done and searchable through expert search. Expert search still fails to check the field types of fields that come from outside the base revision type.

+++ metka/src/main/java/fi/uta/fsd/metka/mvc/domain/ExpertSearchService.java
@@ -46,7 +46,7 @@ public class ExpertSearchService {
             for(RevisionResult result : results) {
                 try {
                     Pair<Boolean, LocalDateTime> info = general.getRevisionableRemovedInfo(result.getId());
-                    RevisionData revision = general.getRevision(result.getId(), result.getNo());
+                    RevisionData revision = general.getRevision(result.getId(), result.getNo().intValue());
                     ExpertSearchRevisionQueryResult qr = new ExpertSearchRevisionQueryResult();
                     if(info.getLeft()) {
                         qr.setState(UIRevisionState.REMOVED);
